requirements.txt
# html
requests
beautifulsoup4
lxml
# db
psycopg2
# other
python-dotenv
pytz